# src/main/java/com/example/bankcards/dto/RoleUpdateDTO.java
package com.example.bankcards.dto;

import jakarta.validation.constraints.NotNull;
import jakarta.validation.constraints.Pattern;
import jakarta.validation.constraints.Size;
import lombok.Getter;
import lombok.Setter;

@Setter
@Getter
public class RoleUpdateDTO {

    @NotNull
    @Size(min = 4, max = 50, message = "Role must be 'USER' or 'ADMIN'")
    @Pattern(regexp = "^(USER|ADMIN)$", message = "Role must be USER or ADMIN")
    private String role;

}