# src/main/java/com/example/bankcards/dto/UserDTO.java
package com.example.bankcards.dto;

import lombok.Data;

import java.util.List;

@Data
public class UserDTO {
    private Long id;
    private String username;
    private String role;
    private List<CardDTO> cards;
}